requirements.txt
pandas
numpy
scikit-learn==1.2.2
tensorflow
joblib
python-dotenv